% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_setup_file.R
\name{blogdown_setup_file}
\alias{blogdown_setup_file}
\alias{makefile_setup_file}
\title{Converts existing setup file to Quarto setup file}
\usage{
blogdown_setup_file(project_folder = here::here(), setup_override = list())

makefile_setup_file(project_folder = here::here(), setup_override = list())
}
\arguments{
\item{project_folder}{Root folder location of the original site}

\item{setup_override}{A list of Quarto settings to be used at conversion}
}
\description{
Converts existing setup file to Quarto setup file
}
