% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{makefile_to_quarto}
\alias{makefile_to_quarto}
\alias{blogdown_to_quarto}
\title{Converts the current site to a Quarto site}
\usage{
makefile_to_quarto(
  project_folder = here::here(),
  quarto_folder = new_quarto_folder(),
  setup_override = list()
)

blogdown_to_quarto(
  project_folder = here::here(),
  quarto_folder = new_quarto_folder(),
  setup_override = list()
)
}
\arguments{
\item{project_folder}{Root folder location of the original site}

\item{quarto_folder}{Target of new Quarto site. It defaults to a new project
folder with the same name as the \code{project_folder}, but with "_quarto" suffixed
to it.}

\item{setup_override}{A list of Quarto settings to be used at conversion}
}
\description{
Converts the current site to a Quarto site
}
