% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_setup_file.R
\name{save_quarto_yaml}
\alias{save_quarto_yaml}
\title{Saves YAML file}
\usage{
save_quarto_yaml(x, location)
}
\arguments{
\item{x}{yaml object}

\item{location}{Path to where to save the new file}
}
\description{
Saves YAML file
}
\details{
It replaces 'yes' with 'true'
}
